Made css changes and improvements
