+++ client/src/stylesheets/footer.css
@@ -1,5 +1,5 @@
 .footer {
-    background-color: rgb(22, 22, 22);;
+    background-color: rgb(22, 22, 22);
     padding: 30px;
 }
