--- client/src/stylesheets/loginscreen.css
@@ -1,64 +1,0 @@
-.loginForm {
-    display: flex;
-    align-content: center;
-    flex-direction: column;
-    background-image: url('/images/splash.JPG');
-    flex: 1;
-    background-size: cover;
-    background-repeat: no-repeat;
-    box-sizing: border-box;
-    align-items: center;
-    justify-content: center;
-    padding: 20em 20em;
-}
-
-.login-username {
-    width:  200px;
-    padding: 10px;
-    margin: auto;
-    margin-bottom:15px;
-} 
-
-.pass {
-    width:  200px;
-    padding: 10px;
-    margin: auto;
-    margin-bottom:30px;
-    margin-top: 15px;
-}
-
-.loginbtn {
-    width: 200px;
-    margin: auto;
-    border-radius: 6px;
-    font-size: 19px;
-    background-color: rgb(59, 154, 213);
-    color: azure;
-    text-shadow:
-   -1px -1px 0 #000,  
-    1px -1px 0 #000,
-    -1px 1px 0 #000,
-    1px 1px 0 #000;
-    font-family: 'Squada One', sans-serif;
-}
-.loginbtn:hover {
-    color: rgb(108, 189, 238);
-    
-}
-.login-header {
-    letter-spacing: 0.02em;
-    font-weight: 550;
-    font-family: 'Squada One', sans-serif;
-    text-align: center;
-    padding: 50px;
-    color: rgb(95, 176, 227);
-    text-shadow:
-   -1px -1px 0 #000,  
-    1px -1px 0 #000,
-    -1px 1px 0 #000,
-    1px 1px 0 #000;
-
-}
-
-
-
