+++ client/src/stylesheets/signupform.css
@@ -49,6 +49,8 @@
     width:  200px;
     font-size: 16px;
     font-family: 'Squada One', sans-serif;
+    background: rgb(0, 0, 0);
+    background: rgba(228, 228, 228, 0.776);
 }
 
 .signup-form-submit {
