+++ client/src/stylesheets/signupscreen.css
@@ -1,7 +1,30 @@
-.signup-header { 
-    margin-top: 35px;
+.sign-main {
+    background-color: rgb(22, 22, 22); 
+}
+
+.sign-container {
+    background-image: url('/images/waterfall.JPG');
+    background-size: cover;
+    background-repeat: no-repeat;
     text-align: center;
-    color:  rgb(59, 154, 213);
+    height: 100%;
+    width: 95%;
+    background-position: center;
+    margin: auto;
+}
+
+.header-container {
+    display: flex;
+    padding-top: 5%;
+    margin-bottom: 30px;
+}
+
+.sign-title { 
+    margin: auto; 
+    box-sizing: border-box;
+    background: rgb(0, 0, 0);
+    background: rgba(245, 238, 238, 0.598);
+    color:  rgb(93, 178, 232);
     font-family: 'Squada One', sans-serif;
     font-weight: 500;
     text-shadow:
@@ -10,4 +33,6 @@
     -1px 1px 0 #000,
     1px 1px 0 #000;
     letter-spacing: 0.05em;
+    padding: 20px;
 }
+
